# apps/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LifeHub",
        "version": "0.1.0.1",
        "description": "API for the Personal Everything App"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/v1/dashboard": {
            "get": {
                "operationId": "v1.dashboard.v1.dashboard",
                "tags": [
                    "Dashboard"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/dashboard/pins/{pin}": {
            "put": {
                "operationId": "v1.dashboard.pins.update",
                "tags": [
                    "Pin"
                ],
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "description": "The pin slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PinUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Pin updated successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.dashboard.pins.destroy",
                "tags": [
                    "Pin"
                ],
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "description": "The pin slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/search/history/{searchHistory}": {
            "delete": {
                "operationId": "v1.search.history.destroy",
                "tags": [
                    "SearchHistory"
                ],
                "parameters": [
                    {
                        "name": "searchHistory",
                        "in": "path",
                        "required": true,
                        "description": "The search history ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/dashboard/search-providers/{searchProvider}": {
            "put": {
                "operationId": "v1.dashboard.search.providers.update",
                "tags": [
                    "SearchProvider"
                ],
                "parameters": [
                    {
                        "name": "searchProvider",
                        "in": "path",
                        "required": true,
                        "description": "The search provider slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchProviderUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Pin updated successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.dashboard.search.providers.destroy",
                "tags": [
                    "SearchProvider"
                ],
                "parameters": [
                    {
                        "name": "searchProvider",
                        "in": "path",
                        "required": true,
                        "description": "The search provider slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 3,
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uri",
                            "maxLength": 2000
                        }
                    },
                    {
                        "name": "term_field",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 10
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 10
                        }
                    },
                    {
                        "name": "icon_color",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "pattern": "^#[0-9a-fA-F]{6}$",
                            "maxLength": 20
                        }
                    },
                    {
                        "name": "default",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Search Provider deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/ai/models/{model}": {
            "get": {
                "operationId": "v1.ai.models.show",
                "tags": [
                    "UserAiModel"
                ],
                "parameters": [
                    {
                        "name": "model",
                        "in": "path",
                        "required": true,
                        "description": "The model ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fields[user_ai_models]",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "provider_id",
                                    "name",
                                    "enabled",
                                    "supports_text",
                                    "supports_images",
                                    "supports_tts",
                                    "supports_stt",
                                    "supports_embeddings",
                                    "supports_reranking",
                                    "created_at",
                                    "updated_at"
                                ]
                            }
                        },
                        "explode": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`UserAiModelResource`",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserAiModelResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "v1.ai.models.update",
                "tags": [
                    "UserAiModel"
                ],
                "parameters": [
                    {
                        "name": "model",
                        "in": "path",
                        "required": true,
                        "description": "The model ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fields[user_ai_models]",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "provider_id",
                                    "name",
                                    "enabled",
                                    "supports_text",
                                    "supports_images",
                                    "supports_tts",
                                    "supports_stt",
                                    "supports_embeddings",
                                    "supports_reranking",
                                    "created_at",
                                    "updated_at"
                                ]
                            }
                        },
                        "explode": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserAiModelUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`UserAiModelResource`",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserAiModelResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.ai.models.destroy",
                "tags": [
                    "UserAiModel"
                ],
                "parameters": [
                    {
                        "name": "model",
                        "in": "path",
                        "required": true,
                        "description": "The model ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PinUpdateRequest": {
                "type": "object",
                "properties": {
                    "section_slug": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "title": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 255
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "icon": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 10
                    },
                    "icon_color": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "pattern": "^#[0-9a-fA-F]{6}$",
                        "maxLength": 20
                    },
                    "order": {
                        "type": "integer"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "tags": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "section_slug",
                    "title",
                    "url",
                    "order",
                    "active"
                ],
                "title": "PinUpdateRequest"
            },
...
            "SearchProviderUpdateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 255
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "maxLength": 2000
                    },
                    "term_field": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "icon": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 10
                    },
                    "icon_color": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "pattern": "^#[0-9a-fA-F]{6}$",
                        "maxLength": 20
                    },
                    "default": {
                        "type": "boolean"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "order": {
                        "type": "integer"
                    }
                },
                "required": [
                    "name",
                    "url",
                    "term_field",
                    "order"
                ],
                "title": "SearchProviderUpdateRequest"
            },
...
            "UserAiModelResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "const": "user_ai_models"
                    },
                    "attributes": {
                        "type": "object",
                        "properties": {
                            "provider_id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "enabled": {
                                "type": "boolean"
                            },
                            "supports_text": {
                                "type": "boolean"
                            },
                            "supports_images": {
                                "type": "boolean"
                            },
                            "supports_tts": {
                                "type": "boolean"
                            },
                            "supports_stt": {
                                "type": "boolean"
                            },
                            "supports_embeddings": {
                                "type": "boolean"
                            },
                            "supports_reranking": {
                                "type": "boolean"
                            },
                            "created_at": {
                                "type": [
                                    "string",
                                    "null"
                                ],
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": [
                                    "string",
                                    "null"
                                ],
                                "format": "date-time"
                            }
                        }
                    }
                },
                "required": [
                    "id",
                    "type"
                ],
                "title": "UserAiModelResource"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}